requirements.txt
openai>=1.0.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
openai-whisper>=20240930
torch>=2.0.0 